WIP clase cerc și planșă
Observație: M_PI este doar în math.h furnizat ca extensie
dependentă de compilator.
De-abia în standardul C++20 pi și alte constante matematice
au fost incluse în standard.

+++ Cerc.h
@@ -13,7 +13,7 @@ public:
     Cerc(int raza);
     ~Cerc();
     friend std::ostream &operator<<(std::ostream &os, const Cerc &cerc);
-    float arie() const;
+    double arie() const;
 };
 
 #endif //FORME_GEOMETRICE_1512_2021_CERC_H
